+++ src/components/tasks/task-view-form/components/descriptionBlock.js
@@ -6,13 +6,7 @@
 export const createDescription = (description) => {
   return (`
     <div class="card__textarea-wrap">
-      <label>
-        <textarea
-          class="card__text"
-          placeholder="Start typing your text here..."
-          name="text"
-        >${description}</textarea>
-      </label>
+      <p class="card__text">${description}</p>
     </div>
   `);
 };
