+++ src/components/tasks/task-view-form/components/settingsBlock.js
@@ -1,7 +1,4 @@
 import {MONTH_NAMES} from "../../../consts";
-import {createDateBlock} from "./dateBlock";
-import {createRepeatBlock} from "./repeatBlock";
-import {createColorsBlock} from "./colorsBlock";
 import {formatTime} from "../../../utils";
 
 /**
@@ -10,8 +7,8 @@ import {formatTime} from "../../../utils";
  * @param {Boolean} isRepeatingTask флаг, определеяющий является ли задача повторяющейся
  * @return {string} разметка блока параметров задачи
  */
-export const createSettingsBlock = (task, isRepeatingTask) => {
-  const {dueDate, color, repeatingDays} = task;
+export const createSettingsBlock = (task) => {
+  const {dueDate} = task;
 
   const isDateShowing = !!dueDate;
   const date = isDateShowing ? `${dueDate.getDate()} ${MONTH_NAMES[dueDate.getMonth()]}` : ``;
@@ -21,11 +18,14 @@ export const createSettingsBlock = (task, isRepeatingTask) => {
     <div class="card__settings">
       <div class="card__details">
         <div class="card__dates">
-          ${createDateBlock(isDateShowing, date, time)}
-          ${createRepeatBlock(isRepeatingTask, repeatingDays)}
+          <div class="card__date-deadline">
+            <p class="card__input-deadline-wrap">
+              <span class="card__date">${date}</span>
+              <span class="card__time">${time}</span>
+            </p>
+          </div>
         </div>
       </div>
-      ${createColorsBlock(color)}
     </div>
   `);
 };
